Display of point cloud bounding boxes
Add an option to display the bounding boxes of all currently loaded
point clouds.

Also included are several tweaks to make the coordinate system more
consistent.  All objects now conceptually live in the global coordinate
system, but are drawn with respect to an offset in order to avoid
numerical precision problems.