Intensity controls via point vertex shader
Implement a vertex shader for points to allow custom dynamic adjustment
of exposure/contrast, and custom point size computation.  This removes
the need for GLEW, so have removed that from the build.  The shaders are
compiled into the binary using the qt resource system to allow the
executable to be completely standalone.