External C++11 bindings for displaz
This is a first go at external C++ bindings for displaz, to enable other
projects to write point clouds to displaz with ease.  There's a few
things which need cleaning up, for example, the ability to specify
vector semantics and to have unscaled integer attributes, but it should
be useful as it stands.