Move to a more robust and simpler draw cost model
Revert to a simple cost model based purely on the number of point cloud
vertices drawn as a function of quality.  This is much easier to make
robust than a multidimensional cost model, due to difficulties fitting
the parameters of the multidimensional model.

Also included various cleanups to comments etc.