Bugfix: Fix camera centre and cursor dragging precision (#33)
Remove QVector3D from various places in the camera code, since the
compoenents are stored in float precision which isn't enough for us.  In
particular, this fixes positioning of the camera centre and dragging of
the 3D cursor so that they're not obviously quantized in the presence of
large coordinate system offsets.