Added support for polygons in Ply files
Add support for drawing polygons from a Ply file.

The format for defining polygons in the Ply file is a `polygon` element with the properties: outer_vertex_index, inner_polygon_vertex_counts, inner_vertex_index.

`outer_vertex_index` is a list of indices of vertices defining the outer ring of the polygon.
`inner_polygon_vertex_counts` is a list containing the number of vertices in each hole of the polygon.
`inner_vertex_index` is a list of all the indices of the vertices composing the inner polygons defining the holes.