Refactor FileLoader thread handling
Move all file loading to a single persistent thread.  If the user spams
displaz with file load requests (easy to do using a script), this avoids
creating a new OS thread for each request.  The amount of work required
to create a thread for each file load seems to have been the underlying
reason why spamming the socket interface led to the GUI becoming
unresponsive.  (See also issue issue #42, which this helps to further
address.)

In addition to the performance improvements, it's actually a lot cleaner
to have moved file loading out of the geometry model class (less lines
FTW!)

Current version has a slight flaw in that the maximum point count isn't
really respected between files, but the previous logic was buggy in this
area in any case.