Make point selection more robust to noise
Point ray origin point closer to cursor so that noise points between the
camera and the point of interest are less likely to be selected.
There's a tradeoff here: points nearer the camera should be selected
where possible, but generally not if they're noise.  We really need a
way to deweight the noise, but this is a stopgap for now.