Move shader handling to Shader menu
Having the actions for both shaders and data sets all together in the
File menu always confused me.  If the author is confused by the UI this
must surely be a bad sign!

All shader handling is now moved into its own menu for reduced
confusion.  The compile action and associated shortcut are discoverable
via the menu, along with adding them to the menu in the shader editor
widget.