More precise closest point algorithm
Efficiently compute the true closest point in the elliptical distance
measure.  Traverses the point cloud octree using a priority queue
ordering, with priority given by the lower bound of the distance to
points in a node.  Combined with early bailout, this very efficiently
finds the exact nearest point in the elliptical distance function.

Fixes #61, since this also works when the user clicks outside the
bounding volume of a point cloud.  Add test data for this case.