Attempt to fix leaking OpenGL state drawing bounding boxes
There was a segfault with Nvidia driver 346.82, inside
PointArray::drawPoints() after reloading a pair of files with the
keyboard shortcut F5 (see #122).  Commenting out the code associated
with initializing the bounding box vertex array inside Geometry.cpp
works around the segfault.  However, it's not clear that this is the
root cause and there was actually a bug in this part of the code.

For now, refactor things to send the bounding box vertices to the GPU
every frame in the hope that this will fix things.  I suppose this is a
bit inefficient, but the main point cloud data is already sent this way.